Make sure Jenkins build fails when there are errors

+++ _scripts/publish-to-github.sh
@@ -14,7 +14,7 @@ rake --trace  gen[production]
 
 rc=$?
 if [[ $rc != 0 ]] ; then
-    echo "ERROR: Execution of Rake script failed."
+    echo "ERROR: Rake script failed!"
     exit $rc
 fi
 
@@ -29,17 +29,32 @@ git fetch origin
 git reset --hard origin/master
 
 ant all.doc
+rc=$?
+if [[ $rc != 0 ]] ; then
+    echo "ERROR: Specification build failed!"
+    exit $rc
+fi
+
 
 # Synchronize the content with the latest-draft folder of the site
 rsync -av --delete --exclude ".git" build/en/html_single/ ../../_site/latest-draft/spec/ 
-
+rc=$?
 popd
+if [[ $rc != 0 ]] ; then
+    echo "ERROR: Latest-draft sync failed!"
+    exit $rc
+fi
 
 pushd _tmp
 # clone hibernate.github.io in _tmp if not present
 if [ ! -d "beanvalidation.github.io" ];
 then
   git clone --depth 1 git@github.com:beanvalidation/beanvalidation.github.io.git
+  rc=$?
+  if [[ $rc != 0 ]] ; then
+    echo "ERROR: beanvalidation/beanvalidation.github.io cannot be cloned!"
+    exit $rc
+  fi
 fi
 
 cd beanvalidation.github.io
@@ -53,9 +68,21 @@ rsync -av \
       --filter "- /cache" --filter "- .git" \
       ../../_site/ .
 
+rc=$?
+if [[ $rc != 0 ]] ; then
+    echo "ERROR: Site sync failed!"
+    exit $rc
+fi
+
 git add -A .
 if git commit -m "Publish generated site";
 then
  git push origin master;
+ rc=$?
 fi
 popd
+if [[ $rc != 0 ]] ; then
+  echo "ERROR: Cannot push on site repository!"
+  exit $rc
+fi
+
